refactor: reorganize validation and status code configurations for CustomPages, introducing new files for improved structure
